Some more refactors
Since this is the refactor PR
